+++ App/connect_four/lib/screens/game_screen/game_screen.dart
@@ -1,5 +1,7 @@
-import 'package:connect_four/Screens/GameScreen/board.dart';
+
 import 'package:flutter/material.dart';
+
+import 'board.dart';
 
 class GameScreen extends StatelessWidget {
   @override
