+++ App/connect_four/lib/screens/welcome/welcome_body.dart
@@ -1,7 +1,9 @@
-import 'package:connect_four/Screens/GameScreen/game_screen.dart';
-import 'package:flutter/material.dart';
-import 'package:connect_four/Screens/Login/login_screen.dart';
+
 import 'package:connect_four/components/rounded_button.dart';
+import 'package:connect_four/screens/game_screen/game_screen.dart';
+import 'package:connect_four/screens/login/login_screen.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 
 class WelcomeBody extends StatelessWidget {
   @override
